feat: display forecast for all the favorite locations
